requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
PyYAML>=6.0
wandb>=0.15.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-image>=0.20.0
opencv-python>=4.7.0
dvc>=3.0.0
transformers>=4.30.0
clip-by-openai>=1.0
torchmetrics>=0.11.0
datasets>=2.12.0
accelerate>=0.20.0
diffusers>=0.18.0
